LineChartの型を追加
