+++ frontend/package.json
@@ -10,6 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@types/recharts": "^1.8.29",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "recharts": "^2.12.5"
